web/requirements.txt
alembic==1.9.4
aniso8601==9.0.1
attrs==22.2.0
black==23.1.0
click==8.1.3
dnspython==2.3.0
flasgger==0.9.5
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
flask-swagger-ui==4.11.1
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
mistune==2.0.5
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.3
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
